Refactor compare utils into its own module
