+++ src/logs/mod.rs
@@ -1,1 +1,2 @@
+pub mod compare;
 pub mod sentry;
